requirements.txt
PyQt6
python-can
pyserial
cantools